directory restructure and rename
