+++ self_interferometry/acquisition/redpitaya/__init__.py
@@ -1,0 +1,7 @@
+"""Red Pitaya hardware interface module."""
+
+from .manager import RedPitayaManager
+from .redpitaya_config import RedPitayaConfig
+from .scpi import scpi
+
+__all__ = ['RedPitayaManager', 'RedPitayaConfig', 'scpi']
