+++ self_interferometry/analysis/barland_cnn.py
