+++ self_interferometry/analysis/lightning_ensemble.py
@@ -4,9 +4,9 @@
 import torch
 from torch import nn
 
-from self_interferometry.signal_analysis.lightning_standard import Standard
-from self_interferometry.signal_analysis.models_cnn import BarlandCNN, BarlandCNNConfig
-from self_interferometry.signal_analysis.models_tcn import TCN, TCNConfig
+from self_interferometry.analysis.lightning_standard import Standard
+from self_interferometry.analysis.barland_cnn import BarlandCNN, BarlandCNNConfig
+from self_interferometry.analysis.models_tcn import TCN, TCNConfig
 
 
 class Ensemble(Standard):
