+++ tests/plot_waveforms.py
@@ -12,8 +12,8 @@ from matplotlib import gridspec
 from numpy.fft import fft, fftfreq
 from scipy import stats
 
-from self_interferometry.redpitaya.coil_driver import CoilDriver
-from self_interferometry.redpitaya.waveform import Waveform
+from self_interferometry.acquisition.simulations.coil_driver import CoilDriver
+from self_interferometry.acquisition.simulations.waveform import Waveform
 
 
 def calculate_fft(
